requirements.txt
filelock==3.18.0
fsspec==2025.3.2
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
pillow==11.2.1
scikit-learn==1.6.1
scipy==1.15.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing_extensions==4.13.2
